% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mfil_abstract}
\alias{mfil_abstract}
\title{mFiL abstract format.}
\usage{
mfil_abstract(...)
}
\arguments{
\item{...}{Optional arguments to be passed to \code{rmarkdown::pdf_document()}.}
}
\description{
Format for creating an abstract for submission to mFiL.
}
\examples{

\dontrun{
library(rmarkdown)
draft("My-abstract.Rmd", template = "mfil_abstract", package = "mfiltemplate")
}
}
